Make FieldOffset covariant with the field type
The documentation states: "A pointer-to-member can be thought of as a
function from `&T` to `&U` with matching lifetimes".

This updates the PhantomData field to better reflect that statement.
Importantly, this removes the `U:` constraints on auto trait impls,
making e.g. `FieldOffset<T, U>: Sync` even if `U: !Sync`.